+++ programs/native/rewards/Cargo.toml
@@ -16,6 +16,7 @@ serde_derive = "1.0.87"
 solana-logger = { path = "../../../logger", version = "0.12.0" }
 solana-metrics = { path = "../../../metrics", version = "0.12.0" }
 solana-sdk = { path = "../../../sdk", version = "0.12.0" }
+solana-rewards-api = { path = "../rewards_api", version = "0.12.0" }
 
 [dev-dependencies]
 solana-vote-program = { path = "../vote", version = "0.12.0" }
@@ -24,4 +25,3 @@ solana-runtime = { path = "../../../runtime", version = "0.12.0" }
 [lib]
 name = "solana_rewards_program"
 crate-type = ["cdylib"]
-
