+++ programs/native/rewards_api/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "solana-rewards-api"
+version = "0.12.0"
+description = "Solana rewards API"
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+edition = "2018"
+
+[dependencies]
+bincode = "1.0.0"
+serde = "1.0.87"
+serde_derive = "1.0.87"
+solana-sdk = { path = "../../../sdk", version = "0.12.0" }
+
+[lib]
+name = "solana_rewards_api"
+crate-type = ["lib"]
