+++ programs/native/rewards_api/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod rewards_instruction;
+pub mod rewards_program;
+pub mod rewards_transaction;
